renaming com.playonlinux.ui.impl.qt in com.playonlinux.qt
