+++ phoenicis-qt/src/main/java/com/playonlinux/qt/mainwindow/toolbar/ToolBar.java
@@ -16,12 +16,12 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package com.playonlinux.ui.impl.qt.mainwindow.toolbar;
+package com.playonlinux.qt.mainwindow.toolbar;
 
 import static com.playonlinux.core.lang.Localisation.translate;
 
-import com.playonlinux.ui.impl.qt.common.ResourceHelper;
-import com.playonlinux.ui.impl.qt.mainwindow.MainWindow;
+import com.playonlinux.qt.common.ResourceHelper;
+import com.playonlinux.qt.mainwindow.MainWindow;
 import com.trolltech.qt.core.QSize;
 import com.trolltech.qt.core.Qt;
 import com.trolltech.qt.gui.QAction;
