+++ phoenicis-qt/src/main/resources/com/playonlinux/qt/common/script/open.png
