+++ phoenicis-qt/src/main/resources/com/playonlinux/qt/common/shortcut/stop.png
