+++ phoenicis-qt/src/main/resources/com/playonlinux/qt/mainwindow/style.css
